requirements.txt
# =============================================================================
# Local Development Dependencies — NYC TLC BI Pipeline
# =============================================================================
# PURPOSE : Install these in your local Python venv for running dbt, writing
#           DAGs, and ad-hoc Snowflake scripting — OUTSIDE of Docker.
#
# USAGE:
#   python -m venv .venv
#   source .venv/bin/activate        # Windows: .venv\Scripts\activate
#   pip install -r requirements.txt
#
# NOTE: This file is for local dev only. The Airflow Docker image uses
#       orchestration/requirements.txt instead.
# =============================================================================


# --- Azure Blob Storage (upload_to_azure.py bootstrap script) ---------------
azure-storage-blob>=12.19.0,<13.0.0

# --- HTTP client (TLC CDN downloads in upload_to_azure.py) ------------------
requests>=2.31.0,<3.0.0

# --- Snowflake connectivity --------------------------------------------------
snowflake-connector-python[pandas]>=3.6.0,<4.0.0
snowflake-sqlalchemy>=1.5.3,<2.0.0

# --- dbt (run models locally with `dbt run`, `dbt test`, `dbt debug`) --------
dbt-core==1.8.7
dbt-snowflake==1.8.4

# --- Data manipulation (ad-hoc scripts & DAG development) -------------------
pandas>=2.1.0,<3.0.0
pyarrow>=16.0.0,<19.0.0        # Parquet read/write; >=16 required for Python 3.13 wheels

# --- Linting & formatting ----------------------------------------------------
ruff>=0.4.0                    # Fast Python linter (replaces flake8 + isort)
black>=24.0.0                  # Code formatter

# --- Testing -----------------------------------------------------------------
pytest>=8.0.0
pytest-cov>=5.0.0
